Changing return type of getWCoord from float to int. This resolves bug id : 2975412 (Bedgraph files are not displayed properly)
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@5572 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ igb/src/com/affymetrix/igb/glyph/GraphGlyph.java
@@ -117,7 +117,7 @@ public final class GraphGlyph extends Glyph {
 		return graf.hasWidth();
 	}
 
-	private float getWCoord(int i) {
+	private int getWCoord(int i) {
 		return ((GraphIntervalSym) graf).getGraphWidthCoord(i);
 	}
